cleaned up source files
