+++ Speculid/ClassicView.swift
@@ -8,20 +8,6 @@
 import SwiftUI
 import SpeculidKit
 
-@propertyWrapper struct UserDefaultsBacked<Value> {
-    let key: String
-    let defaultValue: Value
-    var storage: UserDefaults = .standard
-
-  var wrappedValue: Value {
-      get { UserDefaults.standard.object(forKey: key) as? Value ?? defaultValue }
-      nonmutating set { UserDefaults.standard.set(newValue, forKey: key) }
-  }
-  var projectedValue: Binding<Value> {
-      Binding<Value>( get: { self.wrappedValue }, set: { newValue in self.wrappedValue = newValue } )
-  }
-}
-
 extension UserDefaults {
   
   @objc
@@ -35,77 +21,7 @@ extension UserDefaults {
   }
 }
 
-public class BookmarkURLCollectionObject : ObservableObject {
 
-  //@AppStorage("bookmarks", store: UserDefaults(suiteName: "MLT7M394S7.group.com.brightdigit.Speculid"))
-//  @UserDefaultsBacked(key: "bookmarks", defaultValue: [String : Data](), storage: UserDefaults.shared)
-  
-  static let shared: UserDefaults  = {
-      let combined = UserDefaults(suiteName: "MLT7M394S7.group.com.brightdigit.Speculid")!
-      combined.register(defaults: ["bookmarks" : [String : Data]()])
-      return combined
-  }()
-  
-  @Published var bookmarks  = [URL : URL]()
-  
-  static func saveBookmark(_ url: URL)  {
-    debugPrint("saving bookmark for \(url)")
-    guard let newData =  try? url.bookmarkData(options: .withSecurityScope, includingResourceValuesForKeys: nil, relativeTo: nil) else {
-      return
-    }
-    var bookmarkMap = Self.shared.bookmarks ?? [String : Data]()
-    bookmarkMap[url.path] = newData
-    Self.shared.bookmarks = bookmarkMap
-    
-  }
-  
-  func reset () {
-    Self.shared.bookmarks = [String : Data]()
-  }
-  
-  func isAvailable(basedOn baseURL: URL?, relativePath: String ...) -> Bool {
-    guard let baseURL = baseURL else {
-      return false
-    }
-    
-    var url = baseURL.deletingLastPathComponent()
-    for path in relativePath {
-      url.appendPathComponent(path)
-    }
-    
-    return self.bookmarks[url] != nil
-  }
-  
-  static func transformPath(_ path: String, withBookmarkData bookmarkData: Data) -> (URL, URL)? {
-    
-    var isStale : Bool = false
-    
-     
-    guard let url = try? URL(resolvingBookmarkData: bookmarkData, options: .withSecurityScope, relativeTo: nil, bookmarkDataIsStale: &isStale) else {
-      return nil
-    }
-    
-
-//    if isStale {
-//      DispatchQueue.global().async {
-//        saveBookmark(url)
-//      }
-//
-//    }
-    debugPrint("Adding \(isStale ? "stale" : "fresh") bookmark of \(path) for \(url)")
-    return (URL(fileURLWithPath: path), url)
-  }
-  
-  init () {
-    let bookmarkPublisher = Self.shared.publisher(for: \.bookmarks).compactMap{$0}
-    bookmarkPublisher.map{
-      Dictionary(uniqueKeysWithValues:
-      $0.compactMap(Self.transformPath))
-    }.receive(on: DispatchQueue.main).assign(to: self.$bookmarks)
-  
-    
-  }
-}
 
 
 struct ClassicView: View {
@@ -118,6 +34,8 @@ struct ClassicView: View {
   
   var canBuild : Bool {
     return url != nil &&
+      
+        self.bookmarkCollection.isAvailable(basedOn: self.url, relativePath: self.document.document.assetDirectoryRelativePath) &&
       self.bookmarkCollection.isAvailable(basedOn: self.url, relativePath: self.document.document.assetDirectoryRelativePath, "Contents.json") &&
       self.bookmarkCollection.isAvailable(basedOn: self.url, relativePath: self.document.document.sourceImageRelativePath)
   }
@@ -170,7 +88,7 @@ struct ClassicView: View {
             }
           }, label: {
             HStack{
-              Image(systemName: self.bookmarkCollection.isAvailable(basedOn: self.url, relativePath: self.document.document.assetDirectoryRelativePath, "Contents.json") ? "lock.open.fill" : "lock.fill")
+              Image(systemName: self.bookmarkCollection.isAvailable(basedOn: self.url, relativePath: self.document.document.assetDirectoryRelativePath) ? "lock.open.fill" : "lock.fill")
               Image(systemName: "photo.fill")
               Text(self.document.document.assetDirectoryRelativePath)
             }
@@ -208,27 +126,7 @@ struct ClassicView: View {
         Button("Build") {
           if let url = self.url {
             
-            self.document.build(fromURL: url, onExport: {
-              urls, callback in
-              exportFiles(moving: urls) { (result) in
-                guard let result = result else {
-                  callback(.failure(NSError()))
-                  return
-                }
-                let flatResult = result.flatMap { (sourceUrls)  in
-                  Result{
-                  try sourceUrls.map{
-                    url throws -> (URL,URL)  in
-                    try fileManagement.saveBookmark(url)
-                    let resultURL = try fileManagement.bookmarkURL(fromURL: url)
-                    return (url, resultURL)
-                  }
-                  }
-                }
-                let dictionaryResult = flatResult.map(Dictionary.init(uniqueKeysWithValues:))
-                callback(dictionaryResult)
-              }
-            })
+            self.document.build(fromURL: url)
           }
         }.disabled(!self.canBuild)
